product-service/requirements.txt
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
flask_mysqldb==1.0.1